+++ packages/design-system-web/CHANGELOG.md
@@ -3,7 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.7.3-pr-36.0 (2022-12-16)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
 ## 1.7.2 (2022-12-16)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+## 1.7.1-pr-36.0 (2022-12-15)
 
 **Note:** Version bump only for package @geovistory/design-system-web
 
@@ -142,6 +155,62 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ## 1.6.2-pr-32.0 (2022-10-26)
 
 **Note:** Version bump only for package @geovistory/design-system-web
+
+## 1.6.2-pr-36.6 (2022-12-14)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.6.2-pr-36.5 (2022-12-14)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.6.2-pr-36.4 (2022-12-13)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.6.2-pr-36.3 (2022-12-13)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.6.2-pr-36.2 (2022-12-13)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.6.2-pr-36.1 (2022-12-13)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.6.2-pr-36.0 (2022-12-12)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
 
 ## 1.6.1 (2022-10-14)